# Cloud/contracts/framebase-api-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Framebase Development API",
    "version": "1.2.0",
    "description": "Private development cloud contract for Framebase Phase 4 organization and recovery."
  },
  "paths": {
...
    "/v1/catalog/bootstrap": {
      "get": {
        "summary": "Paginated catalog bootstrap",
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}